pyproject.toml
[project]
name = "gitscope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "apscheduler>=3.11.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "requests>=2.32.5",
]
